test: Restructure library test codes in-source
