--- src/lib/chrome/bookmark.test.ts
@@ -1,6 +1,0 @@
-import { describe, it } from 'vitest';
-import { createBookmarks } from '~/lib/chrome/bookmark';
-
-describe(createBookmarks, () => {
-	it.todo('creates bookmark recursively');
-});
