+++ src/lib/chrome/bookmark.ts
@@ -32,3 +32,13 @@ export async function createBookmarks(parentId: string, tree: TreeNode<Collectio
 		})
 	);
 }
+
+/* c8 ignore start */
+if (import.meta.vitest) {
+	const { describe, it } = import.meta.vitest;
+
+	describe(createBookmarks, () => {
+		it.todo('creates bookmark recursively');
+	});
+}
+/* c8 ignore stop */
