--- src/lib/raindrop/client.test.ts
@@ -1,6 +1,0 @@
-import { describe, it } from 'vitest';
-
-describe('defaultClient', () => {
-	it.todo('request interceptors should');
-	it.todo('response interceptors should');
-});
