--- src/lib/raindrop/filters.test.ts
@@ -1,3 +1,0 @@
-import { it } from 'vitest';
-
-it.todo('nothing to test yet');
