+++ src/lib/raindrop/highlights.ts
@@ -1,0 +1,7 @@
+/* c8 ignore start */
+if (import.meta.vitest) {
+	const { it } = import.meta.vitest;
+
+	it.todo('nothing to test yet');
+}
+/* c8 ignore stop */
