+++ src/lib/raindrop/manager.ts
@@ -7,3 +7,11 @@ export class Manager {
 		this.raindrop = raindrop;
 	}
 }
+
+/* c8 ignore start */
+if (import.meta.vitest) {
+	const { it } = import.meta.vitest;
+
+	it.todo('nothing to test yet');
+}
+/* c8 ignore stop */
