+++ src/lib/raindrop/resource.ts
@@ -9,3 +9,11 @@ export class Resource<R> {
 		this.rawData = rawData;
 	}
 }
+
+/* c8 ignore start */
+if (import.meta.vitest) {
+	const { it } = import.meta.vitest;
+
+	it.todo('nothing to test yet');
+}
+/* c8 ignore stop */
