--- src/lib/raindrop/types.test.ts
@@ -1,6 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/ban-ts-comment */
-// https://betterprogramming.pub/getting-started-testing-types-in-typescript-f64306ec16b
-import { it } from 'vitest';
-
-it('custom type tests run via TS');
