--- src/lib/store.test.ts
@@ -1,21 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import { persisted } from '~/lib/store';
-
-describe('persisted', () => {
-	it('should implement original store interface', async () => {
-		const store = await persisted('test', '');
-
-		// Subscription test
-		let changed = '';
-		store.subscribe((value) => {
-			changed = value;
-		});
-
-		await store.set('hello world');
-		expect(changed).toEqual('hello world');
-
-		// `.update()`
-		await store.update((value) => value + ', done');
-		expect(changed).toEqual('hello world, done');
-	});
-});
