+++ src/lib/types.ts
@@ -13,3 +13,25 @@ export type Integer<N extends number = number> = number extends N
 
 export type PositiveInteger = number;
 export type NonNegativeInteger = number;
+
+/* c8 ignore start */
+if (import.meta.vitest) {
+	const { it } = import.meta.vitest;
+
+	it('custom type tests run via TS');
+
+	{
+		let value: Integer = 0;
+		value = -1;
+		value - 12416427893;
+		value = 1;
+		value = 417489121616378;
+
+		// @ts-expect-error String value not assignable to integer
+		value = '1';
+
+		// @ts-expect-error String value not assignable to integer
+		value = '1.317819';
+	}
+}
+/* c8 ignore stop */
